fixed vertical sprite clipping
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@4124 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/video/v9990/V9990P1Converter.cc
@@ -159,8 +159,8 @@ void V9990P1Converter<Pixel, zoom>::determineVisibleSprites(
 		byte attr = vram->readVRAMP1(spriteInfo + 3);
 
 		if (!(attr & 0x10)) {
-			unsigned spriteY = vram->readVRAMP1(spriteInfo) + 1;
-			unsigned posY = displayY - spriteY;
+			byte spriteY = vram->readVRAMP1(spriteInfo) + 1;
+			byte posY = displayY - spriteY;
 			if (posY < 16) {
 				visibleSprites[index++] = sprite;
 				if (index == 16) break;
